Airflow/requirements.txt
yfinance==0.2.44
pandas==2.1.4
numpy==1.26.4
requests==2.32.3
snowflake-connector-python==3.10.0
apache-airflow-providers-snowflake==5.8.0
dbt-core==1.8.8
dbt-snowflake==1.8.4
dbt-adapters==1.18.0
dbt-common==1.27.1
dbt-protos==1.0.317
